bugfix 306 【需求】权限管理-DNS视图/ACL/转发组-普通用户只读的情况，应该隐藏编辑。删除，新建相关按钮

+++ src/views/dns/forward/ForwardGroup.vue
@@ -73,18 +73,23 @@ export default {
           align: "right",
           width: 160,
           render: (h, { row }) => {
-            return h("div", [
-              h("btn-edit", {
-                on: {
-                  click: () => this.handleOpenEdit(row)
-                }
-              }),
-              h("btn-del", {
-                on: {
-                  click: () => this.handleDelete(row)
-                }
-              })
-            ]);
+
+            if (this.$store.getters.hasPermissionToCreate) {
+              return h("div", [
+                h("btn-edit", {
+                  on: {
+                    click: () => this.handleOpenEdit(row)
+                  }
+                }),
+                h("btn-del", {
+                  on: {
+                    click: () => this.handleDelete(row)
+                  }
+                })
+              ]);
+            } else {
+              return h("div");
+            }
           }
         }
       ],
